init commit-1

+++ src/Ticket.java
@@ -1,11 +1,9 @@
-import java.math.BigDecimal;
-
 public class Ticket {
     private int ticketId;
     private String eventName;
-    private BigDecimal ticketPrice;
+    private double ticketPrice;
 
-    public Ticket(int ticketId, String eventName, BigDecimal ticketPrice) {
+    public Ticket(int ticketId, String eventName, double ticketPrice) {
         this.ticketId = ticketId;
         this.eventName = eventName;
         this.ticketPrice = ticketPrice;
@@ -27,11 +25,11 @@ public class Ticket {
         this.eventName = eventName;
     }
 
-    public BigDecimal getTicketPrice() {
+    public double getTicketPrice() {
         return ticketPrice;
     }
 
-    public void setTicketPrice(BigDecimal ticketPrice) {
+    public void setTicketPrice(Double ticketPrice) {
         this.ticketPrice = ticketPrice;
     }
